refactoring and housekeeping
