+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "firelordjs",
-	"version": "1.4.8",
+	"version": "1.4.9",
 	"description": "🔥 Extremely High Precision Typescript Wrapper for Firestore Web, Providing Unparalleled Type Safe and Dev Experience",
 	"source": "src/index.ts",
 	"main": "dist/src/index.js",
