+++ src/types/index.ts
@@ -16,3 +16,6 @@ export * from './cursor'
 export * from './queryConstraintLimitation'
 export * from './fieldPath'
 export * from './validID'
+export * from './alias'
+export * from './WriteBatch'
+export * from './transaction'
